Address Copilot feedback

+++ dnn-providers/hip-kernel-provider/kernels/RMSnorm/RMSNormFwd.cpp
@@ -1,12 +1,11 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
 // SPDX-License-Identifier: MIT
-#include "CstdInt.hpp"
 #include "FloatTypes.h"
 
-constexpr unsigned LOCAL_SIZE = HIP_PLUGIN_RMSNORM_LOCAL_SIZE;
-constexpr uint64_t C_SIZE = HIP_PLUGIN_RMSNORM_C_SIZE;
-constexpr uint64_t C_STRIDE = HIP_PLUGIN_RMSNORM_C_STRIDE;
-constexpr uint64_t N_STRIDE = C_SIZE * C_STRIDE;
+constexpr unsigned int LOCAL_SIZE = HIP_PLUGIN_RMSNORM_LOCAL_SIZE;
+constexpr size_t C_SIZE = HIP_PLUGIN_RMSNORM_C_SIZE;
+constexpr size_t C_STRIDE = HIP_PLUGIN_RMSNORM_C_STRIDE;
+constexpr size_t N_STRIDE = C_SIZE * C_STRIDE;
 
 using IOType = HIP_PLUGIN_RMSNORM_IO_TYPE;
 
@@ -17,16 +16,16 @@ extern "C" __global__ void RMSnormFwd(const IOType* __restrict__ x,
                                       FLOAT_ACCUM* __restrict__ rstd,
                                       float eps)
 {
-    const uint64_t gid = blockIdx.x;
-    const uint64_t lid = threadIdx.x;
-    const uint64_t o = gid / C_STRIDE;
-    const uint64_t s = gid % C_STRIDE;
+    const unsigned int gid = blockIdx.x;
+    const unsigned int lid = threadIdx.x;
+    const unsigned int o = gid / C_STRIDE;
+    const unsigned int s = gid % C_STRIDE;
 
     FLOAT_ACCUM pvar(0);
     __shared__ FLOAT_ACCUM ltmp[LOCAL_SIZE];
 
     // reduce sum
-    for(uint64_t i = lid; i < C_SIZE; i += LOCAL_SIZE)
+    for(unsigned int i = lid; i < C_SIZE; i += LOCAL_SIZE)
     {
         size_t x_idx = (o * N_STRIDE) + (i * C_STRIDE) + s;
         FLOAT_ACCUM tmp = CVT_FLOAT2ACCUM(x[x_idx]);
@@ -35,7 +34,7 @@ extern "C" __global__ void RMSnormFwd(const IOType* __restrict__ x,
 
     ltmp[lid] = pvar;
     __syncthreads();
-    for(uint32_t i = LOCAL_SIZE >> 1; i > 0; i >>= 1)
+    for(unsigned int i = LOCAL_SIZE >> 1; i > 0; i >>= 1)
     {
         if(lid < i)
         {
@@ -53,7 +52,7 @@ extern "C" __global__ void RMSnormFwd(const IOType* __restrict__ x,
     }
 
     // forward calculation
-    for(uint64_t i = lid; i < C_SIZE; i += LOCAL_SIZE)
+    for(unsigned int i = lid; i < C_SIZE; i += LOCAL_SIZE)
     {
         size_t idx = (o * N_STRIDE) + (i * C_STRIDE) + s;
         FLOAT_ACCUM y_val = (CVT_FLOAT2ACCUM(x[idx])) * prstd * weight[i];
